moving setListeners function from VolumeSlider.kt to VolumeSliderListeners.kt class with companion object
